move on upload images functions to new functions folder
