+++ functions/src/functions/images/on-upload-listing-image.ts
