+++ functions/src/index.ts
@@ -22,5 +22,5 @@ export { default as updateFCMToken } from './user/updateFCMToken';
 export { default as updateUserProfile } from './user/updateUserProfile';
 export { default as createWishlist } from './wishlist/createWishlist';
 export { default as deleteWishlist } from './wishlist/deleteWishlist';
-export { onUploadListingImage } from './images/listings/on-upload-listing-image';
-export { default as uploadProfileImageHandler } from './images/profilePhotos/on-upload-profile-image';
+export { onUploadListingImage } from './functions/images/on-upload-listing-image';
+export { uploadProfileImageHandler } from './functions/images/on-upload-profile-image';
